Reduce code duplication in cell parsing and writing.
