+++ spatialprofilingtoolbox/workflow/defaults/core.py
@@ -12,6 +12,8 @@ import pandas as pd
 from spatialprofilingtoolbox.workflow.defaults.computational_design import ComputationalDesign
 from spatialprofilingtoolbox.workflow.dataset_designs.multiplexed_imaging.halo_cell_metadata_design\
     import HALOCellMetadataDesign
+from spatialprofilingtoolbox.workflow.common.sqlite_context_utility import \
+    WaitingDatabaseContextManager
 from spatialprofilingtoolbox.workflow.common.file_io import raw_line_count
 from spatialprofilingtoolbox.workflow.common.dichotomization import dichotomize
 from spatialprofilingtoolbox.workflow.common.logging.performance_timer import PerformanceTimer
@@ -50,12 +52,42 @@ class CoreJob(ABC):
         self.outcome = outcome
         self.timer = PerformanceTimer()
 
-    @abstractmethod
     def initialize_metrics_database(self):
         """
-        Initialize the local sqlite database for intermediate metrics. The URI is
-        provided by computational_design.get_database_uri() .
+        Setup for the default pipeline-specific database to store intermediate
+        outputs. This method initializes this database's tables.
+        May be overridden in specific workflows.
         """
+        cells_header = self.computational_design.get_cells_header(style='sql')
+        connection, cursor = self.connect_to_intermediate_database()
+        cmd = ' '.join([
+            'CREATE TABLE IF NOT EXISTS',
+            'cells',
+            '(',
+            'id INTEGER PRIMARY KEY AUTOINCREMENT,',
+            ', '.join([' '.join(entry) for entry in cells_header]),
+            ');',
+        ])
+        cursor.execute(cmd)
+        cursor.close()
+        connection.commit()
+        connection.close()
+
+        # Check if fov_lookup is still used
+        fov_lookup_header = self.computational_design.get_fov_lookup_header()
+        connection, cursor = self.connect_to_intermediate_database()
+        cmd = ' '.join([
+            'CREATE TABLE IF NOT EXISTS',
+            'fov_lookup',
+            '(',
+            'id INTEGER PRIMARY KEY AUTOINCREMENT,',
+            ', '.join([' '.join(entry) for entry in fov_lookup_header]),
+            ');',
+        ])
+        cursor.execute(cmd)
+        cursor.close()
+        connection.commit()
+        connection.close()
 
     def connect_to_intermediate_database(self):
         connection = sqlite3.connect(self.computational_design.get_database_uri())
@@ -182,3 +214,113 @@ class CoreJob(ABC):
             logger.debug(
                 'Creating dummy "%s" until its use is fully deprecated.', fov)
             table[fov] = ['FOV1' for i, row in table.iterrows()]
+
+    def create_cell_table(self):
+        """
+        :return:
+            - `cells`. Table of cell data.
+            - `fov_lookup`. FOV descriptor strings in terms of pairs (sample identifier,
+              FOV index integer).
+        :rtype: pandas.DataFrame, dict
+        """
+        cell_groups = []
+        fov_lookup = {}
+        filename = self.input_filename
+        sample_identifier = self.sample_identifier
+        self.timer.record_timepoint('Start reading table')
+        table_file = self.get_table(filename)
+        self.timer.record_timepoint('Finished reading table')
+        self.dataset_design.normalize_fov_descriptors(table_file)
+        self.timer.record_timepoint('Finished normalizing FOV strings in place')
+
+        col = self.dataset_design.get_fov_column()
+        fovs = sorted(list(set(table_file[col])))
+        for i, fov in enumerate(fovs):
+            fov_lookup[(sample_identifier, i)] = fov
+            table_file.loc[table_file[col] == fov, col] = i
+        self.timer.record_timepoint('Finished converting FOVs to integers')
+
+        for _, table_fov in table_file.groupby(col):
+            self.timer.record_timepoint('Start per-FOV cell table parsing')
+            table = table_fov.copy()
+            self.timer.record_timepoint('Finished copying FOV cells table')
+            table = table.reset_index(drop=True)
+            self.timer.record_timepoint('Finished resetting cells table index')
+
+            self.deal_with_compartments(table)
+            phenotype_membership_columns = self.add_and_return_membership_columns(table)
+
+            extra = self.get_and_add_extra_columns(table)
+
+            table['sample_identifier'] = sample_identifier
+            table['outcome_assignment'] = self.outcome
+
+            table = self.restrict_to_pertinent_columns(table, phenotype_membership_columns, extra)
+
+            cell_groups.append(table)
+            self.timer.record_timepoint('Finished parsing one FOV cell table.')
+        logger.debug('%s cells parsed from file %s.', table_file.shape[0], filename)
+        logger.debug('Completed cell table collation.')
+        return pd.concat(cell_groups), fov_lookup
+
+    def get_and_add_extra_columns(self, table):
+        raise NotImplementedError()
+
+    def write_cell_table(self, cells):
+        """
+        Writes cell table to database.
+
+        :param cells: Table of cell areas with sample ID, outcome, etc.
+        :type cells: pandas.DataFrame
+        """
+        self.timer.record_timepoint('Writing parsed cells to file')
+        uri = self.computational_design.get_database_uri()
+        connection = sqlite3.connect(uri)
+        cells.reset_index(drop=True, inplace=True)
+        cells_columns = cells.columns
+        schema_columns = self.computational_design.get_cells_header(
+            style='sql')
+        if all(cells_columns[i] == schema_columns[i][0] for i in range(len(cells_columns))):
+            logger.debug(
+                'Cells table to be written has correct (normalized, ordered) sql-style header '
+                'values.')
+        else:
+            logger.debug(
+                'Cells table to be written has INCORRECT sql-style header values.')
+            if set(cells_columns) == set(schema_columns):
+                logger.debug(
+                    'At least the sets are the same, only the order is wrong.')
+            logger.error('Cannot write cell table with wrong headers.')
+        cells.to_sql('cells', connection, if_exists='append', index_label='id')
+        connection.commit()
+        connection.close()
+        self.timer.record_timepoint('Done writing parsed cells to file')
+
+    def write_fov_lookup_table(self, fov_lookup):
+        """
+        Writes field of view descriptor string lookup table to database.
+
+        :param fov_lookup: Mapping from pairs (sample identifier, FOV index integer) to
+            FOV descriptor strings.
+        :type fov_lookup: dict
+        """
+        keys_list = [
+            column_name for column_name, dtype in self.computational_design.get_fov_lookup_header()
+        ]
+        uri = self.computational_design.get_database_uri()
+        with WaitingDatabaseContextManager(uri) as manager:
+            for (sample_identifier, fov_index), fov_string in fov_lookup.items():
+                values_list = [
+                    '"' + sample_identifier + '"',
+                    str(fov_index),
+                    '"' + fov_string + '"',
+                ]
+                keys = '( ' + ' , '.join(keys_list) + ' )'
+                values = '( ' + ' , '.join(values_list) + ' )'
+                cmd = 'INSERT INTO fov_lookup ' + keys + ' VALUES ' + values + ' ;'
+                try:
+                    manager.execute(cmd)
+                except sqlite3.OperationalError as exception:
+                    logger.error('SQL query failed: %s', cmd)
+                    print(exception)
+        self.timer.record_timepoint('Done writing FOV lookup')
